Sheeren and Farah

+++ restaurants/views.py
@@ -2,6 +2,7 @@ from django.shortcuts import render, redirect
 from .models import Restaurant, Item
 from .forms import RestaurantForm, ItemForm, SignupForm, SigninForm
 from django.contrib.auth import login, authenticate, logout
+from django.contrib import messages
 
 def signup(request):
     form = SignupForm()
@@ -60,6 +61,9 @@ def restaurant_detail(request, restaurant_id):
 
 def restaurant_create(request):
     form = RestaurantForm()
+    if request.user.is_anonymous:
+        messages.warning(request, "Please log in!!")
+        return redirect('signin')
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES)
         if form.is_valid():
@@ -75,6 +79,8 @@ def restaurant_create(request):
 def item_create(request, restaurant_id):
     form = ItemForm()
     restaurant = Restaurant.objects.get(id=restaurant_id)
+    if not (request.user.is_staff or  restaurant.owner== request.user):
+        return redirect('access-denied')
     if request.method == "POST":
         form = ItemForm(request.POST)
         if form.is_valid():
@@ -90,6 +96,10 @@ def item_create(request, restaurant_id):
 
 def restaurant_update(request, restaurant_id):
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
+    if not (request.user.is_staff or restaurant_obj.owner== request.user):
+        return redirect('access-denied')
+        
+    restaurant_obj = Restaurant.objects.get(id=restaurant_id)
     form = RestaurantForm(instance=restaurant_obj)
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES, instance=restaurant_obj)
@@ -102,7 +112,19 @@ def restaurant_update(request, restaurant_id):
     }
     return render(request, 'update.html', context)
 
+
+
 def restaurant_delete(request, restaurant_id):
+    if not (request.user.is_staff):
+        
+        return redirect('access-denied')
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
     restaurant_obj.delete()
     return redirect('restaurant-list')
+
+
+
+
+def access(request):
+ 
+  return render(request, 'access_denied.html')
